+++ src/patch/core/Settings.zig
@@ -114,6 +114,7 @@ pub fn init() void {
     SettingsState.cam7.add("default_rotation_speed", u32, 3);
     SettingsState.cam7.add("default_rotation_smoothing", u32, 0);
     SettingsState.cam7.add("default_planar_movement", bool, false);
+    SettingsState.cam7.add("default_hide_ui", bool, false);
     SettingsState.cam7.add("mouse_dpi", u32, 1600);
     SettingsState.cam7.add("mouse_cm360", f32, 24.0);
     SettingsState.manager.add(&SettingsState.cam7);