+++ src/patch/core/core.zig
@@ -14,3 +14,4 @@ pub const Testing = @import("Testing.zig");
 pub const Allocator = @import("Allocator.zig");
 pub const Debug = @import("Debug.zig");
 pub const Freeze = @import("Freeze.zig");
+pub const HideRaceUI = @import("HideRaceUI.zig");